Common graph test restructure
