+++ dnn-providers/hip-kernel-provider/src/tests/engines/asm_sdpa_engine/GraphTest.hpp
@@ -1,0 +1,28 @@
+// Copyright © Advanced Micro Devices, Inc., or its affiliates.
+// SPDX-License-Identifier:  MIT
+
+#pragma once
+
+#include <flatbuffers/flatbuffers.h>
+#include <hipdnn_data_sdk/flatbuffer_utilities/GraphWrapper.hpp>
+
+#include <memory>
+#include <string>
+#include <utility>
+
+struct GraphTest
+{
+    std::shared_ptr<flatbuffers::DetachedBuffer> buffer;
+    std::string message;
+
+    GraphTest(flatbuffers::FlatBufferBuilder&& builder, std::string inMessage)
+        : buffer(std::make_shared<flatbuffers::DetachedBuffer>(builder.Release()))
+        , message(std::move(inMessage))
+    {
+    }
+
+    hipdnn_data_sdk::flatbuffer_utilities::GraphWrapper graphWrapper() const
+    {
+        return hipdnn_data_sdk::flatbuffer_utilities::GraphWrapper(buffer->data(), buffer->size());
+    }
+};
